+++ plugins/core/jetbrains-community/src/software/aws/toolkits/jetbrains/core/notifications/ProcessNotificationsBase.kt
@@ -3,9 +3,16 @@
 
 package software.aws.toolkits.jetbrains.core.notifications
 
+import com.intellij.notification.NotificationType
+import com.intellij.openapi.actionSystem.AnAction
+import com.intellij.openapi.components.Service
+import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
+import software.aws.toolkits.jetbrains.utils.notifyStickyWithData
 
+@Service(Service.Level.PROJECT)
 class ProcessNotificationsBase {
+    private val notifListener = mutableListOf<NotifListener>()
     init {
         // TODO: install a listener for the polling class
     }
@@ -21,14 +28,41 @@ class ProcessNotificationsBase {
 
     fun processNotification(project: Project, notificationData: NotificationData) {
         val shouldShow = RulesEngine.displayNotification(project, notificationData)
-        if(shouldShow) {
-            // TODO: notifies listeners
+        if (shouldShow) {
+            val notificationContent = notificationData.content.locale
+            val severity = notificationData.severity
+            val followupActions = NotificationManager.createActions(notificationData.actions, notificationContent.description, notificationContent.title)
+            showToast(
+                notificationContent.title,
+                notificationContent.description,
+                NotificationManager.buildNotificationActions(followupActions),
+                checkSeverity(severity),
+                notificationData.id
+            )
+            if (severity == "Critical") {
+                notifyListenerForNotification(notificationContent.title, notificationContent.description, followupActions)
+            }
         }
     }
 
-    fun notifyListenerForNotification() {
+    private fun showToast(title: String, message: String, action: List<AnAction>, notificationType: NotificationSeverity, notificationId: String) {
+        val notifyType = when (notificationType) {
+            NotificationSeverity.CRITICAL -> NotificationType.ERROR
+            NotificationSeverity.WARNING -> NotificationType.WARNING
+            NotificationSeverity.INFO -> NotificationType.INFORMATION
+        }
+        notifyStickyWithData(notifyType, title, message, null, action, notificationId)
     }
 
-    fun addListenerForNotification() {
+    fun notifyListenerForNotification(title: String, description: String, followUpActions: List<NotificationActionList>) =
+        notifListener.forEach { it(title, description, followUpActions) }
+
+    fun addListenerForNotification(newNotifListener: NotifListener) =
+        notifListener.add(newNotifListener)
+
+    companion object {
+        fun getInstance(project: Project): ProcessNotificationsBase = project.service()
     }
 }
+
+typealias NotifListener = (title: String, message: String, followupActions: List<NotificationActionList>) -> Unit
